+++ apps/ledger-live-desktop/src/renderer/modals/ActionConfirmationDialog/index.tsx
@@ -44,7 +44,7 @@ export function ActionConfirmationDialog() {
 
   return (
     <Dialog open={isOpen} onOpenChange={handleOpenChange}>
-      <DialogContent className="bg-canvas rounded-2xl overflow-hidden">
+      <DialogContent className="bg-canvas rounded-2xl overflow-hidden pb-0">
         <DialogHeader appearance="compact" onClose={dismiss} className="absolute right-0 z-10" />
         <div className="relative flex flex-col items-center gap-32 overflow-hidden px-16 pb-24 pt-64">
           <div className="pointer-events-none absolute inset-x-0 top-0 h-full bg-gradient-muted" />
